fix(posts): 修复 Server/Client 组件边界问题导致的 "Element type is invalid" 错误
根本原因:
- PostsListContent 是 async 服务器组件
- 但它直接使用了客户端组件 (Table, Button, Chip 等)
- Next.js 13+ / React 19 不允许服务器组件直接渲染客户端组件

解决方案:
1. 创建 posts-list-client.tsx 作为客户端边界组件
   - 标记为 "use client"
   - 包含所有 UI 渲染逻辑
   - 接收 posts 数据作为 props

2. 修改 page.tsx
   - PostsListContent 保持为服务器组件
   - 仅负责数据获取
   - 将数据传递给客户端组件渲染

此模式与 Gallery 页面保持一致，符合 Next.js 最佳实践。

验证:
- /admin/posts 页面成功加载 (200)
- 不再出现 "Element type is invalid" 错误
- 服务器/客户端组件边界正确分离
